Restructured login route and added readme file.
